+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Fixed individual checkin page
 * Fixed report modal background
 * Fixed venue pages
+* Fixed some dropdown box styles
+* Improved account settings page
 
 
 ### 2021-07-06 - 1.0.0